pyproject.toml
[tool.poetry]
name = "day-21"
version = "0.1.0"
description = ""
authors = ["GabdullinAleksandr <alexrexp@gmail.com>"]
readme = "README.md"
packages = [{include = "day_21"}]

[tool.poetry.dependencies]
python = "^3.8"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
